chore: improve docs
