+++ elements/jsonform/stories/jsonform.stories.js
@@ -89,12 +89,15 @@ export const WKT = WKTStory;
  * JSON Form based on drawtools - Returns the value as GeoJSON
  */
 export const Geojson = GeoJSONStory;
+
+/**
+ * With the `customEditorInterfaces` property it is possible to create
+ * one or more custom form inputs for json-editor (based on its `JSONEditor.AbstractEditor`).
+ * See [json-editor readme](https://github.com/json-editor/json-editor?tab=readme-ov-file#custom-editor-interfaces) for more details.
+ */
+export const CustomEditorInterfaces = CustomEditorInterfacesStory;
+
 /**
  * Unstyled JSON Form
  */
 export const Unstyled = UnStyledStory;
-
-/**
- * Custom editor interfaces for eox-jsonform
- */
-export const CustomEditorInterfaces = CustomEditorInterfacesStory;
